requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
scikit-learn
numpy
pandas
matplotlib
seaborn
pyngrok
